Mark gatherers as polluting, depending on neighbours Wrote pollution system to increment amount polluting This is based on neighbours and will subtract from wallet A number of usize value types changed to signed integers

+++ src/components/power_bar.rs
@@ -1,9 +1,9 @@
 use specs::{Component, VecStorage};
 
-const MAX_POWER: usize = 10_000;
+const MAX_POWER: i32 = 10_000;
 
 pub struct PowerBar {
-    pub power_left: usize,
+    pub power_left: i32,
 }
 
 impl PowerBar {
@@ -17,7 +17,7 @@ impl PowerBar {
         MAX_POWER as f32
     }
 
-    pub fn add_power(&mut self, power: usize) {
+    pub fn add_power(&mut self, power: i32) {
         self.power_left += power;
         if self.power_left > MAX_POWER {
             self.power_left = MAX_POWER;
